refactor: extract room store domain
